Fix sign issue

+++ adelie/src/include/adelie_core/solver/solver_base.hpp
@@ -94,7 +94,7 @@ inline void update_abs_grad(
             try_failed = true;
         }
     };
-    util::omp_parallel_for(routine, 0, groups.size(), n_threads * (n_threads <= groups.size()));
+    util::omp_parallel_for(routine, 0, groups.size(), n_threads * (n_threads <= static_cast<size_t>(groups.size())));
     if (try_failed) {
         throw util::adelie_core_solver_error(
             "exception raised in constraint->solve_zero(). "
